Add back paging and fix a bunch of other things
This ensures that stats are correct even when going fairly far in the
past.

Apparently Octokit.auto_paginate does not work as described, or at all
really.

Also open issues and closed issues are handled correctly such that if an
issue is opened and closed in the same week the numbers should be
open_issues: 1 and closed_issues: 1.